# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Origin FGL HRMS Backend",
    "description": "Phase 2a backend contract for Approval Inbox.",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": ["auth"],
        "summary": "Get Current User",
        "operationId": "get_current_user",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/employees/{employee_id}": {
      "get": {
        "tags": ["employees"],
        "summary": "Get Employee",
        "operationId": "get_employee",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Employee Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/employees/{employee_id}/leave-balance": {
      "get": {
        "tags": ["employees"],
        "summary": "Get Leave Balance",
        "operationId": "get_leave_balance",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Employee Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveBalanceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CurrentUserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "User id."
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "User email."
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "Resolved RBAC role."
          },
          "employee_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Id",
            "description": "Linked employee id, if any."
          }
        },
        "type": "object",
        "required": ["id", "email", "role"],
        "title": "CurrentUserResponse"
      },
      "EmployeeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Prefixed employee id."
          },
          "employee_no": {
            "type": "string",
            "title": "Employee No",
            "description": "Stable employee number."
          },
          "first_name": {
            "type": "string",
            "title": "First Name",
            "description": "Employee first name."
          },
          "last_name": {
            "type": "string",
            "title": "Last Name",
            "description": "Employee last name."
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "Employee work email."
          },
          "department_id": {
            "type": "string",
            "title": "Department Id",
            "description": "Department id."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Job title."
          },
          "manager_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manager Id",
            "description": "Direct manager employee id."
          },
          "employment_type": {
            "$ref": "#/components/schemas/EmploymentType",
            "description": "Employment type."
          },
          "hire_date": {
            "type": "string",
            "format": "date",
            "title": "Hire Date",
            "description": "Hire date."
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "title": "Birth Date",
            "description": "Birth date used only for statutory leave defaults."
          },
          "status": {
            "$ref": "#/components/schemas/EmployeeStatus",
            "description": "Employee lifecycle status."
          },
          "termination_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Termination Date",
            "description": "Termination date if terminated."
          }
        },
        "type": "object",
        "required": [
          "id",
          "employee_no",
          "first_name",
          "last_name",
          "email",
          "department_id",
          "title",
          "employment_type",
          "hire_date",
          "birth_date",
          "status"
        ],
        "title": "EmployeeResponse"
      },
      "EmployeeStatus": {
        "type": "string",
        "enum": ["active", "on_leave", "terminated"],
        "title": "EmployeeStatus"
      },
      "EmploymentType": {
        "type": "string",
        "enum": ["full_time", "part_time", "contractor"],
        "title": "EmploymentType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LeaveBalanceResponse": {
        "properties": {
          "employee_id": {
            "type": "string",
            "title": "Employee Id",
            "description": "Employee id."
          },
          "year": {
            "type": "integer",
            "title": "Year",
            "description": "Balance year."
          },
          "entitled_days": {
            "type": "integer",
            "title": "Entitled Days",
            "description": "Annual entitlement days."
          },
          "carried_over": {
            "type": "integer",
            "title": "Carried Over",
            "description": "Carried-over days."
          },
          "used_days": {
            "type": "integer",
            "title": "Used Days",
            "description": "Approved used days."
          },
          "pending_days": {
            "type": "integer",
            "title": "Pending Days",
            "description": "Pending days."
          },
          "remaining": {
            "type": "integer",
            "title": "Remaining",
            "description": "Computed remaining days."
          }
        },
        "type": "object",
        "required": [
          "employee_id",
          "year",
          "entitled_days",
          "carried_over",
          "used_days",
          "pending_days",
          "remaining"
        ],
        "title": "LeaveBalanceResponse"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email",
            "description": "Demo user email."
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password",
            "description": "Plaintext password submitted over TLS."
          }
        },
        "type": "object",
        "required": ["email", "password"],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "Bearer access token."
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Token type.",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": ["access_token"],
        "title": "LoginResponse"
      },
...
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "RFC 9457 problem type URI.",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short problem title."
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "HTTP status code."
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable problem detail."
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ProblemFieldError"
            },
            "type": "array",
            "title": "Errors",
            "description": "Field-level errors."
          }
        },
        "type": "object",
        "required": ["title", "status", "detail"],
        "title": "ProblemDetail"
      },
      "ProblemFieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Field path associated with the validation or policy error."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable English error message."
          }
        },
        "type": "object",
        "required": ["field", "code", "message"],
        "title": "ProblemFieldError"
      },
...
      "Role": {
        "type": "string",
        "enum": ["employee", "manager", "hr_admin", "executive", "admin"],
        "title": "Role"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}